+++ javascript/array_analysis.js
