+++ tests/capitalize.test.js
@@ -1,4 +1,4 @@
-import capitalize from './capitalize';
+import capitalize from '../javascript/capitalize';
 
 test('hello should equal Hello', () => {
   expect(capitalize('hello')).toBe('Hello');
