+++ tests/reverse-string.test.js
@@ -1,4 +1,4 @@
-import reverseString from './reverse-string';
+import reverseString from '../javascript/reverse-string';
 
 test('world should equal dlrow', () => {
   expect(reverseString('world')).toBe('dlrow');
